Squashed commit of the following:
commit c0f241cdbd898533fc615764bb227d35cbd7fbb8
Author: wode490390 <wode490390@qq.com>
Date:   Fri May 3 17:49:17 2019 +0800

    stuff

+++ src/main/java/cn/nukkit/event/player/PlayerFormRespondedEvent.java
@@ -13,18 +13,18 @@ public class PlayerFormRespondedEvent extends PlayerEvent {
         return handlers;
     }
 
-    protected long formID;
+    protected int formID;
     protected FormWindow window;
 
     protected boolean closed = false;
 
-    public PlayerFormRespondedEvent(Player player, long formID, FormWindow window) {
+    public PlayerFormRespondedEvent(Player player, int formID, FormWindow window) {
         this.player = player;
         this.formID = formID;
         this.window = window;
     }
 
-    public long getFormID() {
+    public int getFormID() {
         return this.formID;
     }
 
@@ -49,5 +49,4 @@ public class PlayerFormRespondedEvent extends PlayerEvent {
     public boolean wasClosed() {
         return window.wasClosed();
     }
-
 }
